+++ internal/pg_documentdb_distributed/src/test/regress/sql/bson_aggregation_pipeline_tests_point_read.sql
@@ -22,11 +22,6 @@ SELECT shard_key_value, object_id, document FROM documentdb_api.collection('agg_
 SELECT document FROM bson_aggregation_find('agg_db', '{ "find": "aggregation_find_point_read", "filter": { "_id": "2" }}');
 EXPLAIN (ANALYZE ON, VERBOSE ON, COSTS ON, BUFFERS OFF, TIMING OFF, SUMMARY OFF) SELECT document FROM bson_aggregation_find('agg_db', '{ "find": "aggregation_find_point_read", "filter": { "_id": "2" }}');
 
--- turned off takes a medium path planner
-SET documentdb.enableFastPathPointLookupPlanner to off;
-EXPLAIN (ANALYZE ON, VERBOSE ON, COSTS ON, BUFFERS OFF, TIMING OFF, SUMMARY OFF) SELECT document FROM bson_aggregation_find('agg_db', '{ "find": "aggregation_find_point_read", "filter": { "_id": "2" }}');
-RESET documentdb.enableFastPathPointLookupPlanner;
-
 -- same point read on non-colocated table - uses slow path
 EXPLAIN (ANALYZE ON, VERBOSE ON, COSTS ON, BUFFERS OFF, TIMING OFF, SUMMARY OFF) SELECT document FROM bson_aggregation_find('db', '{ "find": "aggregation_find_point_read_noncoll", "filter": { "_id": "2" }}');